Improve display of code editor
