+++ less/components/code.less
@@ -45,7 +45,7 @@ pre {
     color: @gb-pre-color;
     background-color: @gb-pre-bg;
     border: 1px solid @gb-pre-border-color;
-    border-radius: @gb-border-radius-base;
+    border-radius: @gb-border-radius-small;
 
     // Account for some code outputs that place code tags in pre tags
     code {
@@ -57,4 +57,3 @@ pre {
         border-radius: 0;
     }
 }
-
