+++ less/variables.less
@@ -356,7 +356,7 @@
 
 @gb-pre-bg:                      #f5f5f5;
 @gb-pre-color:                   #333;
-@gb-pre-border-color:            #ccc;
+@gb-pre-border-color:            @gb-pre-bg;
 
 ///
 /// Tree
